added spec.StringStorage.WalkKeys
